reorganize for modularity
